[CHG] Possibility to specify ingredients to avoid directly in the retriever; Optimized prompt [BUG] set autocommit at connection pool creation

+++ cucinabot/tools.py
@@ -100,7 +100,7 @@ def add(a: float, b: float) -> float:
 
 
 @tool
-def subtract(a: float, b: float) -> int:
+def subtract(a: float, b: float) -> float:
     """
     Subtracts two numbers.
     Args:
